Refactor cni-migration commands to "tools cni-migration"
Move the cni-migration command group to the internal/tools/cnimigration/cmd package and update all references to use the new path. This includes updating the root command registration and documentation.
